Renamed custom_models folder ==> models + Created two separated files for the models - one for user model and one for database model + deleted old migrations folder
