+++ backend/models/database_models.py
@@ -1,23 +1,4 @@
-from django.contrib.auth.models import UserManager, AbstractUser
-from django.db.models import Q
 from django.db import models
-
-##################################################################################
-# USER MODELS
-##################################################################################
-
-
-class CustomUserManager(UserManager):
-    def get_by_natural_key(self, username):
-        return self.get(
-            Q(**{self.model.USERNAME_FIELD: username})
-            | Q(**{self.model.EMAIL_FIELD: username})
-        )
-
-
-class User(AbstractUser):
-    objects = CustomUserManager()
-
 
 ##################################################################################
 # DATABASE MODELS
@@ -35,4 +16,3 @@ class DatabaseCheckModel(models.Model):
     name = models.CharField(max_length=50, blank=False, null=False)
     objects = DatabaseCheckManager()
     REQUIRED_FIELDS = ["name"]
-
